services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic-settings==2.6.1
sqlmodel==0.0.22

boto3>=1.35.70
botocore==1.35.70

playwright==1.49.0
nova-act==0.1.19
